First round of creating Tournament Blueprint
