+++ aesops/__init__.py
@@ -14,3 +14,6 @@ login.login_view = "login"
 
 from aesops import routes
 from data_models import model_store
+
+from .blueprints.tournament_blueprint import tournament_blueprint
+app.register_blueprint(tournament_blueprint)
